refactor: extract compaction_types, progress_display, tool_schema_prune to turn-core
- CompactionTier enum moved from prompts::context to turn-core::compaction_types
- progress_display.rs (474 LOC) extracted to turn-core (all deps already in turn-core)
- tool_schema_prune.rs (466 LOC) extracted to turn-core
- Added PermissionResponseMessaging to orchestration re-exports
- Fixed permission_sync_e2e test import
