+++ rust/crates/astra-turn-core/src/lib.rs
@@ -21,6 +21,7 @@ pub mod chat_history_openai;
 pub mod chat_turn_api_error;
 pub mod chat_turn_explain_wire;
 pub mod chat_turn_heuristics;
+pub mod compaction_types;
 pub mod confidence_contract;
 pub mod context_assembly_trace;
 pub mod edge_executor_id;
@@ -62,6 +63,7 @@ pub mod tool_call_shape;
 pub mod tool_result_sanitize;
 pub mod tool_result_semantics;
 pub mod tool_result_storage;
+pub mod tool_schema_prune;
 pub mod tool_selection;
 pub mod unconsumed;
 pub mod view;
@@ -87,6 +89,7 @@ pub mod headless_tool_journal;
 pub mod orchestration_types;
 pub mod permission_types;
 pub mod persist;
+pub mod progress_display;
 pub mod quality;
 pub mod routing_metrics;
 pub mod tool_registry_chain;
