+++ rust/crates/runtime/tests/permission_sync_e2e.rs
@@ -21,7 +21,8 @@ use astra_runtime::messaging::router::AgentMailboxRouter;
 use astra_runtime::messaging::types::AgentAddress;
 use astra_runtime::orchestration::{
     InheritedPermissions, PermissionDecision, PermissionMode, PermissionRequest,
-    PermissionRequestHandler, PermissionRule, PermissionSyncContext, PermissionUpdate,
+    PermissionRequestHandler, PermissionResponseMessaging, PermissionRule, PermissionSyncContext,
+    PermissionUpdate,
 };
 use astra_runtime::server::delegation_engine::{DelegationTracker, SubRunRecord, SubRunState};
